fix var type

+++ 0x0C-more_malloc_free/1-string_nconcat.c
@@ -11,8 +11,8 @@
 char *string_nconcat(char *s1, char *s2, unsigned int n)
 {
 	char *ptr;
-	int i, size;
-	int j = 0;
+	long unsigned int i, size;
+	long unsigned int j = 0;
 
 	if (s1 == NULL)
 		s1 = "";
